Make overlay dialog more like to detail page.
