+++ app/src/main/res/drawable/ic_reactivex.xml
@@ -1,91 +1,9 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
-  xmlns:aapt="http://schemas.android.com/aapt"
   android:width="24dp"
   android:height="24dp"
-  android:viewportWidth="256"
-  android:viewportHeight="247">
+  android:viewportWidth="24"
+  android:viewportHeight="24">
   <path
-    android:fillColor="#FF0090"
-    android:pathData="M16.69,157.13C3.57,87.38 38.24,20.11 117.81,11.17C106.84,-0.56 91.85,-0.65 84.95,0.78C72.77,4.66 73.15,12.54 59.34,22.72C45.61,30.56 38.71,24.56 28.71,32.69C18.72,40.76 25.69,59.29 21.5,62.89C17.35,70.09 4.46,76.51 2.11,85.61C0.17,97.24 7.37,105.55 7.1,115.53C7.93,123.84 -1.18,128.54 0.13,135.28C4.05,146.28 11.6,152.9 15.16,155.96C15.98,156.53 16.84,157.95 16.69,157.13L16.69,157.13Z" />
-  <path android:pathData="M158.28,60.58C155.37,60.58 153.01,58.22 153.01,55.31C153.01,52.41 155.37,50.05 158.28,50.05C161.18,50.05 163.54,52.41 163.54,55.31C163.54,58.22 161.18,60.58 158.28,60.58L158.28,60.58ZM19.76,164.73C7.15,104.12 46.16,53.47 121.7,78.4C166.03,104.33 221.79,102.65 224.31,85.88C230.51,65.79 196.05,24.38 144.52,13.57C42.28,-6.38 -12.83,104.12 19.76,164.73L19.76,164.73Z">
-    <aapt:attr name="android:fillColor">
-      <gradient
-        android:centerX="178.83"
-        android:centerY="58.38"
-        android:gradientRadius="218.7"
-        android:type="radial">
-        <item
-          android:color="#FFF80090"
-          android:offset="0" />
-        <item
-          android:color="#FF4D008E"
-          android:offset="1" />
-      </gradient>
-    </aapt:attr>
-  </path>
-  <path android:pathData="M187.46,171.49C202.64,173.14 217.05,169.49 230.4,158.61C210.23,181.11 185,192.43 156.06,195.51C169.88,207.25 183.21,212.55 195.96,210.28C160.53,220.06 130.85,209.3 94.74,173.34C92.85,183.02 103.07,198.1 113.61,207.73C52.47,181.22 47.11,98.65 121.67,78.34C44.38,41.21 0.69,113.36 22.16,172.49C43.16,218.03 99.14,253.09 160.49,245.38C190.29,241.73 234.79,221.18 252.97,172.49C240.16,183.98 217.26,194 207.04,194.77C241.63,177.48 260.3,148.1 255.06,107.96C247.9,125.01 238.44,138.11 226.62,147.11C251.74,107.96 247.43,87.72 228.83,65.47C242.13,102.04 224.93,142.63 187.46,171.49L187.46,171.49Z">
-    <aapt:attr name="android:fillColor">
-      <gradient
-        android:centerX="179.14"
-        android:centerY="50.53"
-        android:gradientRadius="212.18"
-        android:type="radial">
-        <item
-          android:color="#FF57008E"
-          android:offset="0" />
-        <item
-          android:color="#FF5C008E"
-          android:offset="0.29" />
-        <item
-          android:color="#FFF80090"
-          android:offset="1" />
-      </gradient>
-    </aapt:attr>
-  </path>
-  <path android:pathData="M169.71,213.63C167.08,213.14 175.66,217.1 159.08,212.71C142.5,208.32 125.62,204.09 94.74,173.34C92.85,183.02 103.07,198.1 113.61,207.73C142.06,227.56 122.71,218.29 166.05,233.27C169.53,226.35 169.71,220.2 169.71,213.63L169.71,213.63Z">
-    <aapt:attr name="android:fillColor">
-      <gradient
-        android:endX="160.06"
-        android:endY="223.88"
-        android:startX="107.95"
-        android:startY="173.34"
-        android:type="linear">
-        <item
-          android:color="#FFF70090"
-          android:offset="0" />
-        <item
-          android:color="#FFE50090"
-          android:offset="0.67" />
-        <item
-          android:color="#33D6008F"
-          android:offset="0.83" />
-        <item
-          android:color="#00C10090"
-          android:offset="1" />
-      </gradient>
-    </aapt:attr>
-  </path>
-  <path android:pathData="M114.6,57.85C114.6,57.85 118.37,52.29 119.84,49.78C121.63,46.71 124.39,41.1 124.39,41.1C124.39,41.1 95.39,31.64 88.27,30.49C66.15,36.23 66.06,45.5 78.44,59.69C79.81,61.27 114.6,57.85 114.6,57.85L114.6,57.85Z">
-    <aapt:attr name="android:fillColor">
-      <gradient
-        android:endX="91.89"
-        android:endY="45.88"
-        android:startX="103.92"
-        android:startY="52.52"
-        android:type="linear">
-        <item
-          android:color="#26B2008F"
-          android:offset="0" />
-        <item
-          android:color="#66F70090"
-          android:offset="0.4" />
-        <item
-          android:color="#E3F60090"
-          android:offset="0.65" />
-        <item
-          android:color="#FFFF0090"
-          android:offset="1" />
-      </gradient>
-    </aapt:attr>
-  </path>
+    android:fillColor="#B7178C"
+    android:pathData="M7.402,2.343c-2.224,0.984 -3.615,2.583 -4.4,3.747 -0.852,1.268 -1.184,2.319 -1.184,2.328 -0.002,0.036 0.005,-0.017 0,0.019 -0.246,0.7 -0.473,1.798 -0.473,1.807 -0.265,1.344 -0.142,2.593 -0.142,2.612l0.029,0.246a9.58,9.58 0,0 1,0.132 -1.012c0.019,-0.133 0.37,-2.016 1.694,-3.511 0.142,-0.218 1.675,-2.233 4.495,-2.233 0.606,0 1.24,0.094 1.874,0.283h0.019c0.038,0.01 0.87,0.313 1.173,0.464 0.284,0.142 0.786,0.36 0.795,0.36h0.02c2.715,1.325 5.062,1.75 5.081,1.76 0.672,0.114 1.278,0.17 1.817,0.17 1.079,0 1.855,-0.236 2.262,-0.69 0.237,-0.26 0.307,-0.559 0.293,-0.814 0.019,-0.985 -0.785,-2.044 -1.467,-2.754a10.868,10.868 0,0 0,-1.542 -1.344c-2.044,-1.514 -4.23,-2.28 -6.52,-2.28 -2.347,0.009 -3.937,0.832 -3.956,0.842zM14.821,5.002c0.284,0 0.511,0.227 0.511,0.51a0.509,0.509 0,0 1,-0.511 0.512,0.509 0.509,0 0,1 -0.511,-0.511c0,-0.284 0.227,-0.511 0.511,-0.511zM8.216,3.062c0.075,0 0.123,0.019 0.142,0.019l3.227,1.003 -0.123,0.255 -1.543,-0.482c-0.548,-0.17 -1.12,-0.351 -1.656,-0.52 -0.01,0 -0.028,-0.01 -0.047,-0.01 -0.02,0 -0.048,0 -0.067,0.01 -0.293,0.075 -1.097,0.359 -1.334,0.785a0.481,0.481 0,0 0,-0.038 0.378c0,0.01 0.17,0.54 0.596,1.088l2.14,-0.104 1.428,-0.066 -0.142,0.293 -3.549,0.17c-0.53,-0.643 -0.738,-1.286 -0.738,-1.286C6.181,3.554 8.083,3.07 8.083,3.07c0.045,-0.001 0.09,-0.006 0.133,-0.01zM19.439,18.809c2.177,-0.946 3.104,-2.584 3.104,-2.584 2.035,-2.772 1.344,-5.697 1.344,-5.697 -1.296,2.82 -2.48,3.596 -2.48,3.596 3.19,-4.854 0.02,-7.788 0.02,-7.788C22.723,9.1 21,12.468 21,12.468c-1.448,3.048 -3.502,4.136 -3.502,4.136 2.29,0.426 3.975,-1.117 3.975,-1.117 -3.284,3.549 -6.842,3.379 -6.842,3.379 1.495,1.675 3.738,1.533 3.738,1.533 -2.934,0.69 -5.688,-0.284 -7.95,-2.167 -0.425,-0.35 -0.832,-0.73 -1.21,-1.136 -0.144,-0.15 -0.28,-0.31 -0.417,-0.464 -0.047,1.751 1.779,3.379 1.779,3.379 -2.29,-0.947 -3.34,-3 -3.34,-3s-1.543,-2.631 -0.426,-5.631c1.107,-2.962 4.419,-3.626 4.504,-3.645 -0.251,-0.117 -0.644,-0.292 -0.814,-0.367 -0.293,-0.142 -1.145,-0.455 -1.145,-0.455a6.397,6.397 0,0 0,-1.798 -0.274c-2.792,0 -4.277,2.11 -4.277,2.11 -1.316,1.477 -1.628,3.388 -1.628,3.388 -0.37,1.827 0.028,3.748 0.028,3.748s0.038,0.151 0.048,0.217c0.303,0.72 0.69,1.43 1.145,2.101 2.46,3.672 6.454,4.94 6.454,4.94 5.914,2.006 9.955,-0.946 9.955,-0.946 3.72,-2.555 4.467,-5.508 4.467,-5.508 -3.01,2.338 -4.306,2.12 -4.306,2.12zM1.307,14.948a12.545,12.545 0,0 1,-0.36 -1.78c-0.01,-0.098 -0.025,-0.207 -0.037,-0.302 0,0 -0.133,-1.287 0.151,-2.697 0,0 0.01,-0.038 0.02,-0.095 0.036,-0.203 0.093,-0.407 0.132,-0.577 0.085,-0.37 0.208,-0.814 0.33,-1.174 0,0 1.307,-4.287 5.726,-6.245 0,0 1.524,-0.805 3.823,-0.862 -0.162,-0.16 -0.349,-0.329 -0.51,-0.435 -1.08,-0.72 -2.688,-0.956 -3.663,0.057 -0.293,0.303 -0.54,0.634 -0.814,0.937 -0.312,0.34 -0.69,0.624 -1.126,0.785 -0.378,0.142 -0.757,0.114 -1.145,0.18 -0.397,0.066 -0.804,0.208 -1.126,0.464 -0.35,0.284 -0.492,0.662 -0.53,1.097 -0.038,0.341 -0.028,0.691 -0.047,1.032 -0.048,1.003 -0.37,1.287 -1.089,1.845 -0.303,0.227 -0.558,0.53 -0.747,0.852 -0.568,1.003 0.34,2.044 0.388,3.057 0.01,0.208 -0.01,0.416 -0.085,0.615 -0.076,0.217 -0.228,0.36 -0.35,0.54 -0.171,0.236 -0.285,0.52 -0.237,0.813 0.047,0.293 0.199,0.568 0.34,0.823 0.275,0.455 0.616,0.862 0.975,1.25 0.012,0.016 0.038,0.04 0.054,0.052l-0.073,-0.232z" />
 </vector>
